Split Electron app functions
